fix bugs

+++ core/src/test/java/com/robin/core/test/service/SysUserService.java
@@ -34,7 +34,7 @@ public class SysUserService extends BaseAnnotationJdbcService<SysUser,Long> {
     @Override
     public Long saveEntity(SysUser user){
         try{
-            return getJdbcDao().createVO(user);
+            return (Long)getJdbcDao().createVO(user);
         }catch (DAOException e) {
             throw new ServiceException(e);
         }
